pyproject.toml
[tool.poetry]
name = "dischatbot"
version = "0.1.0"
description = "A discord chat bot"
authors = ["kosayoda <kieransiek@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7.3"
prompt_toolkit = "^3.0.19"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
python-dotenv = "^0.18.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
